List terraform templates
