+++ pkg/adapters/http.go
@@ -9,8 +9,8 @@ import (
 	"net/url"
 
 	"github.com/go-resty/resty/v2"
+
 	pb "github.com/weaveworks/weave-gitops/pkg/api/profiles"
-	"github.com/weaveworks/weave-gitops/pkg/capi"
 	"github.com/weaveworks/weave-gitops/pkg/clusters"
 	"github.com/weaveworks/weave-gitops/pkg/templates"
 )
@@ -114,16 +114,19 @@ func (c *HTTPClient) Source() string {
 }
 
 // RetrieveTemplates returns the list of all templates from the cluster service.
-func (c *HTTPClient) RetrieveTemplates() ([]capi.Template, error) {
+func (c *HTTPClient) RetrieveTemplates(kind templates.TemplateKind) ([]templates.Template, error) {
 	endpoint := "v1/templates"
 
 	type ListTemplatesResponse struct {
-		Templates []*capi.Template
+		Templates []*templates.Template
 	}
 
 	var templateList ListTemplatesResponse
 	res, err := c.client.R().
 		SetHeader("Accept", "application/json").
+		SetQueryParams(map[string]string{
+			"template_kind": kind.String(),
+		}).
 		SetResult(&templateList).
 		Get(endpoint)
 
@@ -135,9 +138,9 @@ func (c *HTTPClient) RetrieveTemplates() ([]capi.Template, error) {
 		return nil, fmt.Errorf("response status for GET %q was %d", res.Request.URL, res.StatusCode())
 	}
 
-	var ts []capi.Template
+	var ts []templates.Template
 	for _, t := range templateList.Templates {
-		ts = append(ts, capi.Template{
+		ts = append(ts, templates.Template{
 			Name:        t.Name,
 			Provider:    t.Provider,
 			Description: t.Description,
@@ -150,18 +153,19 @@ func (c *HTTPClient) RetrieveTemplates() ([]capi.Template, error) {
 
 // RetrieveTemplatesByProvider returns the list of all templates for a given
 // provider from the cluster service.
-func (c *HTTPClient) RetrieveTemplatesByProvider(provider string) ([]capi.Template, error) {
+func (c *HTTPClient) RetrieveTemplatesByProvider(kind templates.TemplateKind, provider string) ([]templates.Template, error) {
 	endpoint := "v1/templates"
 
 	type ListTemplatesResponse struct {
-		Templates []*capi.Template
+		Templates []*templates.Template
 	}
 
 	var templateList ListTemplatesResponse
 	res, err := c.client.R().
 		SetHeader("Accept", "application/json").
 		SetQueryParams(map[string]string{
-			"provider": provider,
+			"provider":      provider,
+			"template_kind": kind.String(),
 		}).
 		SetResult(&templateList).
 		Get(endpoint)
@@ -174,9 +178,9 @@ func (c *HTTPClient) RetrieveTemplatesByProvider(provider string) ([]capi.Templa
 		return nil, fmt.Errorf("response status for GET %q was %d", res.Request.URL, res.StatusCode())
 	}
 
-	var ts []capi.Template
+	var ts []templates.Template
 	for _, t := range templateList.Templates {
-		ts = append(ts, capi.Template{
+		ts = append(ts, templates.Template{
 			Name:        t.Name,
 			Provider:    t.Provider,
 			Description: t.Description,
@@ -188,18 +192,19 @@ func (c *HTTPClient) RetrieveTemplatesByProvider(provider string) ([]capi.Templa
 
 // RetrieveTemplateParameters returns the list of all parameters of the
 // specified template.
-func (c *HTTPClient) RetrieveTemplateParameters(name string) ([]capi.TemplateParameter, error) {
+func (c *HTTPClient) RetrieveTemplateParameters(kind templates.TemplateKind, name string) ([]templates.TemplateParameter, error) {
 	endpoint := "v1/templates/{name}/params"
 
 	type ListTemplateParametersResponse struct {
-		Parameters []*capi.TemplateParameter
+		Parameters []*templates.TemplateParameter
 	}
 
 	var templateParametersList ListTemplateParametersResponse
 	res, err := c.client.R().
 		SetHeader("Accept", "application/json").
 		SetPathParams(map[string]string{
-			"name": name,
+			"name":          name,
+			"template_kind": kind.String(),
 		}).
 		SetResult(&templateParametersList).
 		Get(endpoint)
@@ -212,9 +217,9 @@ func (c *HTTPClient) RetrieveTemplateParameters(name string) ([]capi.TemplatePar
 		return nil, fmt.Errorf("response status for GET %q was %d", res.Request.URL, res.StatusCode())
 	}
 
-	var tps []capi.TemplateParameter
+	var tps []templates.TemplateParameter
 	for _, p := range templateParametersList.Parameters {
-		tps = append(tps, capi.TemplateParameter{
+		tps = append(tps, templates.TemplateParameter{
 			Name:        p.Name,
 			Description: p.Description,
 			Required:    p.Required,
@@ -227,13 +232,13 @@ func (c *HTTPClient) RetrieveTemplateParameters(name string) ([]capi.TemplatePar
 
 // POST request payload
 type TemplateParameterValuesAndCredentials struct {
-	Values      map[string]string `json:"values"`
-	Credentials capi.Credentials  `json:"credentials"`
+	Values      map[string]string     `json:"values"`
+	Credentials templates.Credentials `json:"credentials"`
 }
 
 // RenderTemplateWithParameters returns a YAML representation of the specified
 // template populated with the supplied parameters.
-func (c *HTTPClient) RenderTemplateWithParameters(name string, parameters map[string]string, creds capi.Credentials) (string, error) {
+func (c *HTTPClient) RenderTemplateWithParameters(kind templates.TemplateKind, name string, parameters map[string]string, creds templates.Credentials) (string, error) {
 	endpoint := "v1/templates/{name}/render"
 
 	// POST response payload
@@ -248,7 +253,8 @@ func (c *HTTPClient) RenderTemplateWithParameters(name string, parameters map[st
 	res, err := c.client.R().
 		SetHeader("Accept", "application/json").
 		SetPathParams(map[string]string{
-			"name": name,
+			"name":          name,
+			"template_kind": kind.String(),
 		}).
 		SetBody(TemplateParameterValuesAndCredentials{Values: parameters, Credentials: creds}).
 		SetResult(&renderedTemplate).
@@ -275,16 +281,16 @@ func (c *HTTPClient) RenderTemplateWithParameters(name string, parameters map[st
 func (c *HTTPClient) CreatePullRequestFromTemplate(params templates.CreatePullRequestFromTemplateParams) (string, error) {
 	// POST request payload
 	type CreatePullRequestFromTemplateRequest struct {
-		RepositoryURL   string               `json:"repositoryUrl"`
-		HeadBranch      string               `json:"headBranch"`
-		BaseBranch      string               `json:"baseBranch"`
-		Title           string               `json:"title"`
-		Description     string               `json:"description"`
-		TemplateName    string               `json:"templateName"`
-		ParameterValues map[string]string    `json:"parameter_values"`
-		CommitMessage   string               `json:"commitMessage"`
-		Credentials     capi.Credentials     `json:"credentials"`
-		ProfileValues   []capi.ProfileValues `json:"profile_values"`
+		RepositoryURL   string                    `json:"repositoryUrl"`
+		HeadBranch      string                    `json:"headBranch"`
+		BaseBranch      string                    `json:"baseBranch"`
+		Title           string                    `json:"title"`
+		Description     string                    `json:"description"`
+		TemplateName    string                    `json:"templateName"`
+		ParameterValues map[string]string         `json:"parameter_values"`
+		CommitMessage   string                    `json:"commitMessage"`
+		Credentials     templates.Credentials     `json:"credentials"`
+		ProfileValues   []templates.ProfileValues `json:"profile_values"`
 	}
 
 	// POST response payload
@@ -299,7 +305,7 @@ func (c *HTTPClient) CreatePullRequestFromTemplate(params templates.CreatePullRe
 	)
 
 	endpoint = "v1/clusters"
-	if params.TemplateKind == templates.GitOpsTemplateKind {
+	if params.TemplateKind == templates.GitOpsTemplateKind.String() {
 		endpoint = "v1/tfcontrollers"
 	}
 
@@ -338,11 +344,11 @@ func (c *HTTPClient) CreatePullRequestFromTemplate(params templates.CreatePullRe
 }
 
 // RetrieveCredentials returns a list of all CAPI credentials.
-func (c *HTTPClient) RetrieveCredentials() ([]capi.Credentials, error) {
+func (c *HTTPClient) RetrieveCredentials() ([]templates.Credentials, error) {
 	endpoint := "v1/credentials"
 
 	type ListCredentialsResponse struct {
-		Credentials []*capi.Credentials
+		Credentials []*templates.Credentials
 		Total       int32
 	}
 
@@ -361,9 +367,9 @@ func (c *HTTPClient) RetrieveCredentials() ([]capi.Credentials, error) {
 		return nil, fmt.Errorf("response status for GET %q was %d", res.Request.URL, res.StatusCode())
 	}
 
-	var creds []capi.Credentials
+	var creds []templates.Credentials
 	for _, c := range credentialsList.Credentials {
-		creds = append(creds, capi.Credentials{
+		creds = append(creds, templates.Credentials{
 			Group:     c.Group,
 			Version:   c.Version,
 			Kind:      c.Kind,
@@ -376,8 +382,8 @@ func (c *HTTPClient) RetrieveCredentials() ([]capi.Credentials, error) {
 }
 
 // RetrieveCredentialsByName returns a specific set of CAPI credentials.
-func (c *HTTPClient) RetrieveCredentialsByName(name string) (capi.Credentials, error) {
-	var creds capi.Credentials
+func (c *HTTPClient) RetrieveCredentialsByName(name string) (templates.Credentials, error) {
+	var creds templates.Credentials
 
 	credsList, err := c.RetrieveCredentials()
 	if err != nil {
@@ -386,7 +392,7 @@ func (c *HTTPClient) RetrieveCredentialsByName(name string) (capi.Credentials, e
 
 	for _, c := range credsList {
 		if c.Name == name {
-			creds = capi.Credentials{
+			creds = templates.Credentials{
 				Group:     c.Group,
 				Version:   c.Version,
 				Kind:      c.Kind,
@@ -497,14 +503,14 @@ func (c *HTTPClient) DeleteClusters(params clusters.DeleteClustersParams) (strin
 	endpoint := "v1/clusters"
 
 	type DeleteClustersPullRequestRequest struct {
-		RepositoryUrl string           `json:"repositoryUrl"`
-		HeadBranch    string           `json:"headBranch"`
-		BaseBranch    string           `json:"baseBranch"`
-		Title         string           `json:"title"`
-		Description   string           `json:"description"`
-		ClusterNames  []string         `json:"clusterNames"`
-		CommitMessage string           `json:"commitMessage"`
-		Credentials   capi.Credentials `json:"credentials"`
+		RepositoryUrl string                `json:"repositoryUrl"`
+		HeadBranch    string                `json:"headBranch"`
+		BaseBranch    string                `json:"baseBranch"`
+		Title         string                `json:"title"`
+		Description   string                `json:"description"`
+		ClusterNames  []string              `json:"clusterNames"`
+		CommitMessage string                `json:"commitMessage"`
+		Credentials   templates.Credentials `json:"credentials"`
 	}
 
 	type DeleteClustersResponse struct {
@@ -547,11 +553,11 @@ func (c *HTTPClient) DeleteClusters(params clusters.DeleteClustersParams) (strin
 
 // RetrieveTemplateProfiles returns the list of all profiles of the
 // specified template.
-func (c *HTTPClient) RetrieveTemplateProfiles(name string) ([]capi.Profile, error) {
+func (c *HTTPClient) RetrieveTemplateProfiles(name string) ([]templates.Profile, error) {
 	endpoint := "v1/templates/{name}/profiles"
 
 	type ListTemplatePResponse struct {
-		Profiles []*capi.Profile
+		Profiles []*templates.Profile
 	}
 
 	var templateProfilesList ListTemplatePResponse
@@ -571,9 +577,9 @@ func (c *HTTPClient) RetrieveTemplateProfiles(name string) ([]capi.Profile, erro
 		return nil, fmt.Errorf("response status for GET %q was %d", res.Request.URL, res.StatusCode())
 	}
 
-	var tps []capi.Profile
+	var tps []templates.Profile
 	for _, p := range templateProfilesList.Profiles {
-		tps = append(tps, capi.Profile{
+		tps = append(tps, templates.Profile{
 			Name:              p.Name,
 			Home:              p.Home,
 			Sources:           p.Sources,
